backend/pyproject.toml
[project]
name = "enmusiquer-backend"
version = "0.1.0"
description = "FastAPI backend for Enmusiquer"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116.0,<1.0.0",
  "uvicorn[standard]>=0.35.0,<1.0.0",
  "sqlalchemy>=2.0.41,<3.0.0",
  "psycopg[binary]>=3.2.9,<4.0.0",
  "alembic>=1.16.4,<2.0.0",
  "pydantic-settings>=2.10.1,<3.0.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app"]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0,<9.0.0",
  "httpx>=0.28.0,<1.0.0",
]
